+++ src/bird.c
@@ -74,8 +74,8 @@ void bird_draw(const bird_t bird)
 
 inline static void bird_tick_animation(bird_t *bird)
 {
-    const u64 ticks_ms = get_ticks_ms();
-    u64 anim_ms = bird->anim_ms;
+    const u32 ticks_ms = get_ticks_ms();
+    u32 anim_ms = bird->anim_ms;
     u8 anim_frame = bird->anim_frame;
     switch ( bird->state )
     {
@@ -119,7 +119,7 @@ inline static void bird_tick_sine_wave(bird_t *bird)
     /* Center the bird in the sky */
     bird->y = 0.0;
     /* Periodically update the "floating" effect */
-    const u64 ticks_ms = get_ticks_ms();
+    const u32 ticks_ms = get_ticks_ms();
     if ( ticks_ms - bird->sine_ms >= BIRD_SINE_RATE )
     {
         bird_tick_dx( bird );
@@ -143,7 +143,7 @@ static void bird_tick_velocity(bird_t *bird, const gamepad_state_t gamepad)
         bird->anim_frame = BIRD_ANIM_FRAMES - 1;
         audio_play_sfx( g_audio, SFX_WING );
     }
-    const u64 ticks_ms = get_ticks_ms();
+    const u32 ticks_ms = get_ticks_ms();
     if ( ticks_ms - bird->dy_ms >= BIRD_VELOCITY_RATE )
     {
         bird_tick_dx( bird );
@@ -182,7 +182,7 @@ inline static bird_color_t bird_random_color_type(void)
 
 void bird_tick(bird_t *bird, const gamepad_state_t gamepad)
 {
-    const u64 ticks_ms = get_ticks_ms();
+    const u32 ticks_ms = get_ticks_ms();
     /* State transitions based on button input */
     switch (bird->state)
     {
